Replace old drug grouper with new one
