+++ app/jobs/api_updater.rb
@@ -1,19 +1,22 @@
 class ApiUpdater < Updater
+  attr_reader :updater
+
   def perform()
     import
     pre_grouper = PreGrouper.new
     pre_grouper.perform
     grouper = Grouper.new
-    grouper.perform(should_group_genes?, should_group_drugs?)
+    grouper.perform(should_group_genes?, should_group_drugs?, updater.source.id)
     post_grouper = PostGrouper.new
     post_grouper.perform(should_cleanup_gene_claims?)
   end
 
   def import
+    @updater = create_updater
     updater.update
   end
 
-  def updater
-    raise StandardError.new('Must implement #updater in subclass')
+  def create_updater
+    raise StandardError.new('Must implement #create_updater in subclass')
   end
 end
