+++ app/jobs/update_ckb.rb
@@ -1,5 +1,5 @@
 class UpdateCkb < ApiUpdater
-  def updater
+  def create_updater
     Genome::OnlineUpdaters::Ckb::Updater.new()
   end
